Moving tests out of aer executor into run quantum program
